requirements.txt
pillow>=10.0.0
pyyaml>=6.0.1
numpy>=1.26.0
scipy>=1.11.0
opencv-python>=4.8.0
streamlit>=1.36.0
torch
